backend/requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0

# WebSocket & Channels
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# File handling
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
whitenoise==6.6.0

